Correcciones de las pantallas asociadas al envío y recepción delos pedidos

+++ Toppick_App/lib/Products/Models/especialidad.dart
@@ -16,7 +16,7 @@ class Especialidad extends Producto{
       json['descripcion'],
       json['precio'],
       json['tiempoPreparacion'],
-      (json['calificacion']==null)?0:json['calificacion'],
+      (json['calificacion']==null)?0:json['calificacion'].toDouble(),
       (json['urlImagen']==null)?"assets/img/pescadito.jpg":json['urlImagen'],
       json['categoria'],
       json['tipo']);
